fix(specialmaps): renaming tpzcylindermap files
