--- SpecialMaps/TPZCylinder.hpp
@@ -1,13 +1,0 @@
-//
-//  TPZCylinder.hpp
-//  pz
-//
-//  Created by Philippe Devloo on 04/05/18.
-//
-
-#ifndef TPZCylinder_hpp
-#define TPZCylinder_hpp
-
-#include <stdio.h>
-
-#endif /* TPZCylinder_hpp */
